pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "umap-osm"
version = "3.0.0"
description = "Draw beautiful maps from OpenStreetMap with a simple Python API and CLI."
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [{name = "Umutcan Edizaslan"}]
dependencies = [
    "geopandas",
    "matplotlib",
    "osmnx",
    "psutil",
    "PyYAML",
    "numpy",
    "shapely"
]

[tool.setuptools]
packages = ["umap", "umap.core", "umap.utils"]

[project.urls]
Homepage = "https://github.com/U-C4N/Umap"

[project.scripts]
umap = "umap.cli:main"

[project.optional-dependencies]
dev = ["build", "twine", "pytest"]

[tool.setuptools.package-data]
umap = ["*.yaml", "*.yml"]
